+++ web/app/routes/authenticated/dashboard.ts
@@ -34,6 +34,7 @@ export default class DashboardRoute extends Route {
 
     if (docsAwaitingReview.length > 0) {
       // load owner information
+      debugger;
       await this.store.maybeFetchPeople.perform(docsAwaitingReview);
     }
 
@@ -54,6 +55,7 @@ export default class DashboardRoute extends Route {
      *
      */
     if (this.recentDocs.all) {
+      // FIXME: will this also grab the avatars?
       void this.recentDocs.fetchAll.perform();
     } else {
       await this.recentDocs.fetchAll.perform();
