+++ web/tests/acceptance/application-test.ts
@@ -7,8 +7,10 @@ import {
 } from "ember-simple-auth/test-support";
 import { MirageTestContext, setupMirage } from "ember-cli-mirage/test-support";
 import SessionService from "hermes/services/session";
-import { TEST_SUPPORT_URL } from "hermes/utils/hermes-urls";
-import { authenticateTestUser } from "hermes/utils/mirage-utils";
+import {
+  authenticateTestUser,
+  TEST_SUPPORT_URL,
+} from "hermes/mirage/mirage-utils";
 
 module("Acceptance | application", function (hooks) {
   setupApplicationTest(hooks);
