+++ web/tests/acceptance/authenticated/document-test.ts
@@ -21,12 +21,12 @@ import {
 } from "hermes/components/document/sidebar";
 import { capitalize } from "@ember/string";
 import window from "ember-window-mock";
-import { TEST_SHORT_LINK_BASE_URL } from "hermes/utils/hermes-urls";
 import {
   TEST_USER_2_EMAIL,
   TEST_USER_3_EMAIL,
   TEST_USER_EMAIL,
-} from "hermes/utils/mirage-utils";
+  TEST_SHORT_LINK_BASE_URL,
+} from "hermes/mirage/mirage-utils";
 
 const ADD_RELATED_RESOURCE_BUTTON_SELECTOR =
   "[data-test-section-header-button-for='Related resources']";
