+++ web/tests/integration/components/product/avatar-test.ts
@@ -2,9 +2,7 @@ import { find, render } from "@ember/test-helpers";
 import { hbs } from "ember-cli-htmlbars";
 import { MirageTestContext, setupMirage } from "ember-cli-mirage/test-support";
 import { setupRenderingTest } from "ember-qunit";
-import ConfigService from "hermes/services/config";
 import { setupProductIndex } from "hermes/tests/mirage-helpers/utils";
-import { setFeatureFlag } from "hermes/utils/mirage-utils";
 import { module, test } from "qunit";
 
 const AVATAR = "[data-test-product-avatar]";
